style: move speech recognizer to stt dir
