+++ app/infrastructure/clients/stt/naver/naver_stt_client.py
@@ -2,7 +2,7 @@ from fastapi import HTTPException
 from typing import Dict, Any
 import requests
 
-from app.domain.services.speech_recognizer import SpeechRecognizer
+from app.infrastructure.clients.stt.speech_recognizer import SpeechRecognizer
 from app.common.utils.logger import get_logger
 
 logger = get_logger(__name__)
