+++ app/infrastructure/clients/stt/speech_recognizer.py
